This is the compiled version of this Python tool
Very straightforward usage - enter your current BPM and the tool will output two numbers, a fast bpm and a slow bpm. 'Fast' would be the BPM you are transitioning to (meaning if you create a 3/4 loop at your current BPM, Fast will be the resulting BPM). 'Slow' is the exact opposite, this would be the BPM you are transitioning from (so if you have a known BPM that you want to transition to, Slow is what BPM you need to start at with your 3/4 loop.)